Implement "trimming" of query names
The function does not actually trim the query names, but simply replaces
all but the last two labels with "___". The length corresponds to the
original length, but the exact value is hidden.
Ideally the number of preserved labels should be configurable, but this
does not seem too easy within the excluded_fields.conf